refactor: move feature flipping to engine
