+++ engine/api/observability/observability.go
@@ -11,8 +11,8 @@ import (
 	"go.opencensus.io/trace"
 
 	"github.com/ovh/cds/engine/api/cache"
+	"github.com/ovh/cds/engine/featureflipping"
 	"github.com/ovh/cds/engine/gorpmapper"
-	"github.com/ovh/cds/sdk/featureflipping"
 	"github.com/ovh/cds/sdk/log"
 	"github.com/ovh/cds/sdk/telemetry"
 )
