+++ engine/featureflipping/featureflipping.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ovh/cds/sdk"
 	"github.com/ovh/cds/sdk/log"
 	"github.com/ovh/cds/sdk/luascript"
-  gocache "github.com/patrickmn/go-cache"
+	gocache "github.com/patrickmn/go-cache"
 
 	"github.com/ovh/cds/engine/gorpmapper"
 )
