--- sdk/featureflipping/package.go
@@ -1,2 +1,0 @@
-package featureflipping
-// Doc
